Refactor : Uplode -> Upload 함수 및 변수명 오기입 수정
