+++ src/pages/EditPage/index.tsx
@@ -16,7 +16,7 @@ import {
 
 import MainTemplate from '../../components/template/MainTemplate';
 
-import * as S from '../UplodePage/styles';
+import * as S from '../UploadPage/styles';
 import Input from '../../components/Form/Input';
 import Button from '../../components/Button';
 import Textarea from '../../components/Form/Textarea';
